requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
anyio==4.3.0
asgiref==3.7.2
astunparse==1.6.3
betterpath==0.2.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
Django==4.2.7
docx2txt==0.8
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httplib2==0.22.0
idna==3.6
imageio==2.34.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
l==0.11.0
lazy_loader==0.3
libclang==16.0.6
looseversion==1.3.0
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
pillow==10.2.0
plum-dispatch==1.7.4
protobuf==4.25.3
prov==2.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydot==2.0.0
PyMuPDF==1.23.23
PyMuPDFb==1.23.22
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
Spire.Doc==12.2.0
sqlparse==0.4.4
starlette==0.37.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
traits==6.3.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
vcversioner==2.16.0.0
Werkzeug==3.0.1
wrapt==1.14.1
zope.interface==6.2
